+++ python/cudf/dataframe/dataframe.py
@@ -862,6 +862,87 @@ class DataFrame(object):
         for k in self.columns:
             self[k] = self[k].set_index(idx)
 
+    def reindex(self, labels=None, axis=0, index=None, columns=None,
+                copy=True):
+        """Return a new DataFrame whose axes conform to a new index
+
+        ``DataFrame.reindex`` supports two calling conventions
+        * ``(index=index_labels, columns=column_names)``
+        * ``(labels, axis={0 or 'index', 1 or 'columns'})``
+
+        Parameters
+        ----------
+        labels : Index, Series-convertible, optional, default None
+        axis : {0 or 'index', 1 or 'columns'}, optional, default 0
+        index : Index, Series-convertible, optional, default None
+            Shorthand for ``df.reindex(labels=index_labels, axis=0)``
+        columns : array-like, optional, default None
+            Shorthand for ``df.reindex(labels=column_names, axis=1)``
+        copy : boolean, optional, default True
+
+        Returns
+        -------
+        A DataFrame whose axes conform to the new index(es)
+
+        Examples
+        --------
+        >>> import cudf
+        >>> df = cudf.DataFrame()
+        >>> df['key'] = [0, 1, 2, 3, 4]
+        >>> df['val'] = [float(i + 10) for i in range(5)]
+        >>> df_new = df.reindex(index=[0, 3, 4, 5],
+                                columns=['key', 'val', 'sum'])
+        >>> print(df)
+           key   val
+        0    0  10.0
+        1    1  11.0
+        2    2  12.0
+        3    3  13.0
+        4    4  14.0
+        >>> print(df_new)
+           key   val  sum
+        0    0  10.0  NaN
+        3    3  13.0  NaN
+        4    4  14.0  NaN
+        5   -1   NaN  NaN
+        """
+
+        if labels is None and index is None and columns is None:
+            return self.copy(deep=copy)
+
+        df = self
+        cols = columns
+        dtypes = OrderedDict(df.dtypes)
+        idx = labels if index is None and axis in (0, 'index') else index
+        cols = labels if cols is None and axis in (1, 'columns') else cols
+        df = df if cols is None else df[list(set(df.columns) & set(cols))]
+
+        if idx is not None:
+            idx = idx if isinstance(idx, Index) else as_index(idx)
+            if df.index.dtype != idx.dtype:
+                cols = cols if cols is not None else list(df.columns)
+                df = DataFrame()
+            else:
+                df = DataFrame(None, idx).join(df, how='left', sort=True)
+                # double-argsort to map back from sorted to unsorted positions
+                df = df.take(idx.argsort(True).argsort(True).to_gpu_array())
+
+        idx = idx if idx is not None else df.index
+        names = cols if cols is not None else list(df.columns)
+
+        length = len(idx)
+        cols = OrderedDict()
+
+        for name in names:
+            if name in df:
+                cols[name] = df[name].copy(deep=copy)
+            else:
+                dtype = dtypes.get(name, np.float64)
+                col = columnops.column_empty(length, dtype, True)
+                cols[name] = Series(data=col, index=idx)
+
+        return DataFrame(cols, idx)
+
     def set_index(self, index):
         """Return a new DataFrame with a new index
 
@@ -911,7 +992,6 @@ class DataFrame(object):
 
         result_cols = cpp_copying.apply_gather(cols, positions)
 
-        out = DataFrame()
         for i, col_name in enumerate(self._cols):
             out[col_name] = result_cols[i]
 
@@ -1753,9 +1833,11 @@ class DataFrame(object):
 
         # We save the original categories for the reconstruction of the
         # final data frame
+        categorical_dtypes = {}
         col_with_categories = {}
         for name, col in itertools.chain(lhs._cols.items(), rhs._cols.items()):
             if pd.api.types.is_categorical_dtype(col):
+                categorical_dtypes[name] = col.dtype
                 col_with_categories[name] = col.cat.categories
 
         # Save the order of the original column names for preservation later
@@ -1827,7 +1909,7 @@ class DataFrame(object):
                     mask = Buffer(valid)
                 df[name] = columnops.build_column(
                     Buffer(col),
-                    dtype=col.dtype,
+                    dtype=categorical_dtypes.get(name, col.dtype),
                     mask=mask,
                     categories=col_with_categories.get(name, None),
                 )